requirements.txt
numpy
scikit-learn
tensorflow
pandas
joblib
spacy
ollama
beautifulsoup4
fastapi
uvicorn
PyMuPDF
PyPDF2
nltk
tinydb
bertopic
tf-keras
matplotlib
pytesseract
pdf2image
Pillow